Move mocks so external packages can use them
