+++ confirm/confirm_test.go
@@ -11,7 +11,7 @@ import (
 	"testing"
 
 	"github.com/volatiletech/authboss"
-	"github.com/volatiletech/authboss/internal/mocks"
+	"github.com/volatiletech/authboss/mocks"
 )
 
 func TestInit(t *testing.T) {
