+++ defaults/responder_test.go
@@ -9,7 +9,7 @@ import (
 	"testing"
 
 	"github.com/volatiletech/authboss"
-	"github.com/volatiletech/authboss/internal/mocks"
+	"github.com/volatiletech/authboss/mocks"
 )
 
 type testRenderer struct {
