+++ expire/expire_test.go
@@ -8,7 +8,7 @@ import (
 	"time"
 
 	"github.com/volatiletech/authboss"
-	"github.com/volatiletech/authboss/internal/mocks"
+	"github.com/volatiletech/authboss/mocks"
 )
 
 func TestExpireIsExpired(t *testing.T) {
