+++ otp/twofactor/twofactor_verify_test.go
@@ -8,7 +8,7 @@ import (
 	"testing"
 
 	"github.com/volatiletech/authboss"
-	"github.com/volatiletech/authboss/internal/mocks"
+	"github.com/volatiletech/authboss/mocks"
 )
 
 func TestSetupEmailVerify(t *testing.T) {
