+++ recover/recover_test.go
@@ -13,7 +13,7 @@ import (
 	"time"
 
 	"github.com/volatiletech/authboss"
-	"github.com/volatiletech/authboss/internal/mocks"
+	"github.com/volatiletech/authboss/mocks"
 )
 
 const (
