+++ remember/remember_test.go
@@ -10,7 +10,7 @@ import (
 	"testing"
 
 	"github.com/volatiletech/authboss"
-	"github.com/volatiletech/authboss/internal/mocks"
+	"github.com/volatiletech/authboss/mocks"
 )
 
 func TestInit(t *testing.T) {
